requirements.txt
# Core
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.0

# LLM
anthropic==0.34.0
openai>=1.0.0

# Vector DB & Embeddings
chromadb==0.5.0
sentence-transformers==3.0.0

# Web fetching (for KB building)
httpx==0.27.0
beautifulsoup4==4.12.0
lxml==5.2.0

# Text processing
ebooklib==0.18
nltk==3.8.1

# GraphRAG
networkx==3.3
